--- a/.gitignore
+++ b/.gitignore
@@ -181,3 +181,4 @@ _infra
 
 # dev env
 .devcontainer.json
+tmp