--- a/notebooks/run_florence2sam2_image.ipynb
+++ b/notebooks/run_florence2sam2_image.ipynb
@@ -2,7 +2,7 @@
  "cells": [
   {
    "cell_type": "code",
-   "execution_count": 21,
+   "execution_count": null,
    "id": "6a9968fc-11b3-4513-b277-67869a4d99f1",
    "metadata": {
     "scrolled": true
@@ -29,7 +29,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 22,
+   "execution_count": 2,
    "id": "8339e1e0-e5e6-4fa3-8310-ceb3834184dc",
    "metadata": {},
    "outputs": [],
@@ -219,7 +219,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 23,
+   "execution_count": 3,
    "id": "98f8b038-8d52-4665-821e-5fa6d407d1af",
    "metadata": {},
    "outputs": [],
@@ -233,7 +233,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 24,
+   "execution_count": null,
    "id": "a16a92ae-d17a-46e5-89d9-e4fa8de4ef20",
    "metadata": {},
    "outputs": [],
@@ -267,7 +267,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 26,
+   "execution_count": 6,
    "id": "78134b8b-35fa-4428-b125-16c6a94057ca",
    "metadata": {
     "scrolled": true
@@ -284,7 +284,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 27,
+   "execution_count": 7,
    "id": "aa3e48b9-03cc-4cc4-bfaa-ee4a5f882550",
    "metadata": {},
    "outputs": [],