--- a/notebooks/run_florence2sam2_image.ipynb
+++ b/notebooks/run_florence2sam2_image.ipynb
@@ -258,6 +258,16 @@
    ]
   },
   {
+   "cell_type": "code",
+   "execution_count": null,
+   "id": "272dfb56",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "results[0]"
+   ]
+  },
+  {
    "cell_type": "markdown",
    "id": "fe932d1e-e995-4fbd-8c88-c6075c54cf24",
    "metadata": {},
@@ -267,24 +277,26 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 6,
+   "execution_count": 14,
    "id": "78134b8b-35fa-4428-b125-16c6a94057ca",
    "metadata": {
     "scrolled": true
    },
    "outputs": [],
    "source": [
-    "idx = 0\n",
     "return_frame_data = []\n",
-    "for inner_idx in range(len(results[idx][\"labels\"])):\n",
-    "    mask = rle_decode_array(results[idx][\"masks\"][inner_idx])\n",
-    "    label = str(inner_idx) + \": \" + results[idx][\"labels\"][inner_idx]\n",
-    "    return_frame_data.append({\"label\": label, \"mask\": mask, \"score\": 1.0})"
+    "for frame_idx in range(len(results)):\n",
+    "    annotations = len(results[frame_idx])\n",
+    "    for annotation_idx in range(annotations):\n",
+    "        annotation = results[frame_idx][annotation_idx]\n",
+    "        mask = rle_decode_array(annotation[\"mask\"])\n",
+    "        label = str(annotation[\"id\"]) + \": \" + annotation[\"label\"]\n",
+    "        return_frame_data.append({\"label\": label, \"mask\": mask, \"score\": 1.0})"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": 15,
    "id": "aa3e48b9-03cc-4cc4-bfaa-ee4a5f882550",
    "metadata": {},
    "outputs": [],