--- a/notebooks/run_florence2sam2_video.ipynb
+++ b/notebooks/run_florence2sam2_video.ipynb
@@ -37,7 +37,7 @@
   {
    "cell_type": "code",
    "execution_count": 2,
-   "id": "f62ec322-1692-4a5e-b9aa-140a59c9dd06",
+   "id": "8339e1e0-e5e6-4fa3-8310-ceb3834184dc",
    "metadata": {},
    "outputs": [],
    "source": [
@@ -185,16 +185,8 @@
     "                    draw.rectangle((x, y, text_box[2], text_box[3]), fill=color[label])\n",
     "                    draw.text((x, y), text, fill=\"black\", font=font)\n",
     "        frame_out.append(np.array(pil_image))\n",
-    "    return frame_out[0] if len(frame_out) == 1 else frame_out\n"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 3,
-   "id": "8339e1e0-e5e6-4fa3-8310-ceb3834184dc",
-   "metadata": {},
-   "outputs": [],
-   "source": [
+    "    return frame_out[0] if len(frame_out) == 1 else frame_out\n",
+    "\n",
     "def load_video(video_path: str):\n",
     "    with open(video_path, \"rb\") as f:\n",
     "        video_bytes = f.read()\n",
@@ -214,7 +206,7 @@
     "            cap.release()\n",
     "            return np.array(frames), fps\n",
     "\n",
-    "def rle_decode_array(rle: Dict[str, List[int]]) -> np.ndarray:\n",
+    "def rle_decode_array(rle: dict[str, list[int]]) -> np.ndarray:\n",
     "    r\"\"\"Decode a run-length encoded mask. Returns numpy array, 1 - mask, 0 - background.\n",
     "\n",
     "    Parameters:\n",
@@ -365,7 +357,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 4,
+   "execution_count": 3,
    "id": "98f8b038-8d52-4665-821e-5fa6d407d1af",
    "metadata": {},
    "outputs": [],
@@ -400,7 +392,7 @@
     "start = time.time()\n",
     "results = florence2_sam2(video=frames, prompt=\"shark, person\")\n",
     "end = time.time()\n",
-    "print(f\"Time taken: {end - start} seconds\")\n"
+    "print(f\"Time taken: {end - start} seconds\")"
    ]
   },
   {
@@ -413,7 +405,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": 6,
    "id": "78134b8b-35fa-4428-b125-16c6a94057ca",
    "metadata": {
     "scrolled": true
@@ -433,7 +425,7 @@
   {
    "cell_type": "code",
    "execution_count": 8,
-   "id": "aa3e48b9-03cc-4cc4-bfaa-ee4a5f882550",
+   "id": "39148eb3-267a-47b0-ab2c-33e93f12ae49",
    "metadata": {},
    "outputs": [],
    "source": [
@@ -447,7 +439,7 @@
   {
    "cell_type": "code",
    "execution_count": null,
-   "id": "39148eb3-267a-47b0-ab2c-33e93f12ae49",
+   "id": "a8bef6e0-ff06-41b4-9e61-5a387d732d9e",
    "metadata": {},
    "outputs": [],
    "source": [