--- a/notebooks/run_florence2sam2_video.ipynb
+++ b/notebooks/run_florence2sam2_video.ipynb
@@ -424,7 +424,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 8,
+   "execution_count": 7,
    "id": "39148eb3-267a-47b0-ab2c-33e93f12ae49",
    "metadata": {},
    "outputs": [],