--- a/notebooks/run_florence2sam2_video.ipynb
+++ b/notebooks/run_florence2sam2_video.ipynb
@@ -363,7 +363,8 @@
    "outputs": [],
    "source": [
     "# Path to your video\n",
-    "video_path = \"../tests/shared_data/videos/shark_10fps.mp4\"\n",
+    "# video_path = \"../tests/shared_data/videos/shark_10fps.mp4\"\n",
+    "video_path = \"../tests/shared_data/videos/tracking_car.mp4\"\n",
     "\n",
     "# Load the video\n",
     "frames, fps = load_video(video_path)"
@@ -390,7 +391,8 @@
    "outputs": [],
    "source": [
     "start = time.time()\n",
-    "results = florence2_sam2(video=frames, prompt=\"shark, person\")\n",
+    "# results = florence2_sam2(video=frames, prompt=\"shark, person\")\n",
+    "results = florence2_sam2(video=frames, prompt=\"car\")\n",
     "end = time.time()\n",
     "print(f\"Time taken: {end - start} seconds\")"
    ]
@@ -424,7 +426,17 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": null,
+   "id": "d4cfe190",
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "len(return_data[0])"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
    "id": "39148eb3-267a-47b0-ab2c-33e93f12ae49",
    "metadata": {},
    "outputs": [],
@@ -444,7 +456,8 @@
    "outputs": [],
    "source": [
     "os.makedirs(\"../tmp\", exist_ok=True)\n",
-    "filepath = \"../tmp/florence2sam2_shark_fps_10.mp4\"\n",
+    "# filepath = \"../tmp/florence2sam2_shark_fps_10.mp4\"\n",
+    "filepath = \"../tmp/florence2sam2_track_car.mp44\"\n",
     "output_path = save_video(processed_frames, filepath, fps=int(fps))"
    ]
   },