--- a/notebooks/run_florence2sam2_video.ipynb
+++ b/notebooks/run_florence2sam2_video.ipynb
@@ -363,8 +363,8 @@
    "outputs": [],
    "source": [
     "# Path to your video\n",
-    "# video_path = \"../tests/shared_data/videos/shark_10fps.mp4\"\n",
-    "video_path = \"../tests/shared_data/videos/tracking_car.mp4\"\n",
+    "video_path = \"../tests/shared_data/videos/shark_10fps.mp4\"\n",
+    "# video_path = \"../tests/shared_data/videos/tracking_car.mp4\"\n",
     "\n",
     "# Load the video\n",
     "frames, fps = load_video(video_path)"
@@ -391,8 +391,8 @@
    "outputs": [],
    "source": [
     "start = time.time()\n",
-    "# results = florence2_sam2(video=frames, prompt=\"shark, person\")\n",
-    "results = florence2_sam2(video=frames, prompt=\"car\")\n",
+    "results = florence2_sam2(video=frames, prompt=\"shark, person\")\n",
+    "# results = florence2_sam2(video=frames, prompt=\"car\")\n",
     "end = time.time()\n",
     "print(f\"Time taken: {end - start} seconds\")"
    ]
@@ -415,28 +415,20 @@
    "outputs": [],
    "source": [
     "return_data = []\n",
-    "for idx, frame_i in enumerate(frames):\n",
+    "for frame_idx in range(len(results)):\n",
     "    return_frame_data = []\n",
-    "    for inner_idx in range(len(results[idx][\"labels\"])):\n",
-    "        mask = rle_decode_array(results[idx][\"masks\"][inner_idx])\n",
-    "        label = str(inner_idx) + \": \" + results[idx][\"labels\"][inner_idx]\n",
+    "    annotations = len(results[frame_idx])\n",
+    "    for annotation_idx in range(annotations):\n",
+    "        annotation = results[frame_idx][annotation_idx]\n",
+    "        mask = rle_decode_array(annotation[\"mask\"])\n",
+    "        label = str(annotation[\"id\"]) + \": \" + annotation[\"label\"]\n",
     "        return_frame_data.append({\"label\": label, \"mask\": mask, \"score\": 1.0})\n",
     "    return_data.append(return_frame_data)"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": null,
-   "id": "d4cfe190",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "len(return_data[0])"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 7,
    "id": "39148eb3-267a-47b0-ab2c-33e93f12ae49",
    "metadata": {},
    "outputs": [],
@@ -456,8 +448,8 @@
    "outputs": [],
    "source": [
     "os.makedirs(\"../tmp\", exist_ok=True)\n",
-    "# filepath = \"../tmp/florence2sam2_shark_fps_10.mp4\"\n",
-    "filepath = \"../tmp/florence2sam2_track_car.mp44\"\n",
+    "filepath = \"../tmp/florence2sam2_shark.mp4\"\n",
+    "# filepath = \"../tmp/florence2sam2_track_car.mp4\"\n",
     "output_path = save_video(processed_frames, filepath, fps=int(fps))"
    ]
   },