--- a/vision_agent_tools/helpers/filters.py
+++ b/vision_agent_tools/helpers/filters.py
@@ -10,8 +10,8 @@ _AREA_THRESHOLD = 0.82
 def filter_bbox_predictions(
     predictions: dict[str, Any],
     image_size: tuple[int, int],
-    nms_threshold: float,
     *,
+    nms_threshold: float = 1.0,
     bboxes_key: str = "bboxes",
     label_key: str = "labels",
 ) -> dict[str, Any]: