--- a/vision_agent_tools/models/florence2_qa.py
+++ b/vision_agent_tools/models/florence2_qa.py
@@ -1,9 +1,11 @@
+from typing import Any
+
 import torch
 from PIL import Image
 
 from vision_agent_tools.shared_types import PromptTask
 from vision_agent_tools.models.florence2 import Florence2
-from vision_agent_tools.helpers.roberta_qa import RobertaQA
+from vision_agent_tools.models.roberta_qa import RobertaQA
 from vision_agent_tools.shared_types import BaseMLModel, Device
 
 
@@ -23,7 +25,7 @@ class FlorenceQA(BaseMLModel):
         self._roberta_qa = RobertaQA()
 
     @torch.inference_mode()
-    def __call__(self, image: Image.Image, question: str) -> str:
+    def __call__(self, image: Image.Image, question: str) -> dict[str, Any]:
         """
         FlorenceQA model answers questions about images.
 
@@ -36,12 +38,9 @@ class FlorenceQA(BaseMLModel):
         """
         image = image.convert("RGB")
         task = PromptTask.MORE_DETAILED_CAPTION
-        output_caption = self._florence(image=image, task=task)
-        caption = output_caption[task]
-        answer = self._roberta_qa(caption, question)
-
-        return answer.answer
+        output_caption = self._florence(images=[image], task=task)[0]
+        roberta_pred = self._roberta_qa(output_caption["text"], question)
+        return {"text": roberta_pred["answer"]}
 
     def to(self, device: Device):
-        self._florence.to(device=device)
-        self._roberta_qa.to(device=device)
+        raise NotImplementedError("This method is not supported for FlorenceQA model.")