--- a/vision_agent_tools/models/model_registry.py
+++ b/vision_agent_tools/models/model_registry.py
@@ -1,6 +1,8 @@
 import re
 from typing import Dict, Type
+
 from pydantic import BaseModel, field_validator
+
 from vision_agent_tools.shared_types import BaseMLModel
 
 MODELS_PATH = "vision_agent_tools.models"