… .github/workflows/unit-tests.yml
… README.md
… docs/florence2-qa.md
… docs/florence2-sam2.md
… docs/florence2.md
… docs/florencev2.md
… docs/owlv2.md
… notebooks/run_florence2sam2.ipynb
… poetry.lock
… pyproject.toml
… tests/conftest.py
… tests/helpers/test_filters.py
… tests/models/conftest.py
… tests/models/data/florence2sam2_image_results.json
… tests/models/data/florence2sam2_video_results.json
… tests/models/data/owlv2_video_results.json
… tests/models/florence2/conftest.py
… tests/models/florence2/data/results/caption_to_phrase_grounding_video_results.json
… tests/models/florence2/data/results/ocr_results.json
… tests/models/florence2/test_api.py
… tests/models/florence2/test_caption.py
… tests/models/florence2/test_caption_to_phrase_grounding.py
… tests/models/florence2/test_dense_region_caption.py
… tests/models/florence2/test_detailed_caption.py
… tests/models/florence2/test_more_detailed_caption.py
… tests/models/florence2/test_ocr.py
… tests/models/florence2/test_ocr_with_region.py
… tests/models/florence2/test_od.py
… tests/models/florence2/test_open_vocabulary_detection.py
… tests/models/florence2/test_referring_expression_segmentation.py